Outsource iterator in iter module
