+++ lib/entry/libimagentrycategory/src/store.rs
@@ -33,6 +33,7 @@ use error::CategoryErrorKind as CEK;
 use error::CategoryError as CE;
 use error::ResultExt;
 use error::Result;
+use iter::CategoryNameIter;
 
 pub const CATEGORY_REGISTER_NAME_FIELD_PATH : &'static str = "category.register.name";
 
@@ -219,55 +220,5 @@ fn represents_category(store: &Store, sid: StoreId, name: &str) -> Result<bool> 
                 Ok(bl) // false
             }
         })
-}
-
-/// Iterator for Category names
-///
-/// Iterates over Result<Category>
-///
-/// # Return values
-///
-/// In each iteration, a Option<Result<Category>> is returned. Error kinds are as follows:
-///
-/// * CategoryErrorKind::StoreReadError if a name could not be fetched from the store
-/// * CategoryErrorKind::HeaderReadError if the header of the fetched item couldn't be read
-/// * CategoryErrorKind::TypeError if the name could not be fetched because it is not a String
-///
-pub struct CategoryNameIter<'a>(&'a Store, StoreIdIterator);
-
-impl<'a> CategoryNameIter<'a> {
-
-    fn new(store: &'a Store, sidit: StoreIdIterator) -> CategoryNameIter<'a> {
-        CategoryNameIter(store, sidit)
-    }
-
-}
-
-impl<'a> Iterator for CategoryNameIter<'a> {
-    type Item = Result<Category>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // TODO: Optimize me with lazy_static
-        let query = CATEGORY_REGISTER_NAME_FIELD_PATH;
-
-        while let Some(sid) = self.1.next() {
-            if sid.is_in_collection(&["category"]) {
-                let func = |store: &Store| { // hack for returning Some(Result<_, _>)
-                    store
-                        .get(sid)?
-                        .ok_or_else(|| CE::from_kind(CEK::StoreReadError))?
-                        .get_header()
-                        .read_string(query)
-                        .chain_err(|| CEK::HeaderReadError)?
-                        .map(Category::from)
-                        .ok_or_else(|| CE::from_kind(CEK::StoreReadError))
-                };
-
-                return Some(func(&self.0))
-            } // else continue
-        }
-
-        None
-    }
 }
